requirements.txt
altair
folium
geopy
matplotlib
pillow
requests
streamlit
torch
torchvision